requirements.txt
ultralytics==8.2.80
opencv-python==4.11.0.86
matplotlib==3.10.7
pandas==2.3.3
numpy==1.26.4
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.2
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
psutil==7.1.1
py-cpuinfo==9.0.0
seaborn==0.13.2
ultralytics-thop==2.0.17
kagglehub==0.1.7